pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "notion_telegram_llm"
version = "0.1.0"
description = "Notion + Telegram + LLM task/learning management"
requires-python = ">=3.9"
dependencies = [
    "notion-client==2.2.1",
    "python-telegram-bot==20.8",
    "openai==1.12.0",
    "pydantic==2.6.3",
    "python-dotenv==1.0.1",
    "aiohttp==3.12.7",
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]

[project.optional-dependencies]
dev = [
    "pytest==8.0.2",
    "pytest-asyncio==0.23.5",
    "pytest-cov==4.1.0",
]

[tool.pytest.ini_options]
testpaths = ["src/tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --tb=short"
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "e2e: End-to-end tests",
    "slow: Slow running tests",
] 